% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{data.coor}
\alias{data.coor}
\title{Generate irregular data}
\usage{
data.coor(n, mu.c = 10, nb.n = 10, nb.p = 0.25, mu.f, phi)
}
\arguments{
\item{n}{the number of samples}

\item{mu.c}{the mean}

\item{nb.n}{the maximum number of observations per curve}

\item{nb.p}{the probability of success when computing the number of observations
per curve}
}
\value{
A list with two components:
\item{x}{A list with the \code{n} vectors of observations (one per curve)}
\item{pp}{A corresponding list with the vectors of times at which observations
were taken}
}
\description{
Generate sparse irregular longitudinal samples.
}
\details{
Generate sparse irregular longitudinal samples. The number of observations
per curve is \code{rbinom(nb.n, nb.p)}. It uses a Karhunen-Loeve 
representations uing the mean function \code{mu.f} and (2)
eigenfunctions (in \code{phi}). The scores are \code{rnorm(0, 25)}
and \code{rnorm(0, 1)}.
}
