% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{data.coor2}
\alias{data.coor2}
\title{Generate regular data}
\usage{
data.coor2(n, mu.c = 10, np = 100, mu.f, phi)
}
\arguments{
\item{n}{the number of samples}

\item{mu.c}{the mean}

\item{np}{the number of observations per curve}

\item{mu.f}{the mean function}

\item{phi}{a list of length 2 containing two "eigenfunctions"}
}
\value{
A list with two components:
\item{x}{A list with the \code{n} vectors of observations (one per curve)}
\item{pp}{A corresponding list with the vectors of times at which observations
were taken}
}
\description{
Generate regular longitudinal samples.
}
\details{
Generate regular longitudinal samples. The observations
per curve are \code{seq(0, 1, length=np)}. It uses a Karhunen-Loeve
representations uing the mean function \code{mu.f} and (2)
eigenfunctions (in \code{phi}). The scores are \code{rnorm(0, 25)}
and \code{rnorm(0, 1)}.
}
