% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{k.epan}
\alias{k.epan}
\title{Epanechnikov kernel}
\usage{
k.epan(x)
}
\arguments{
\item{x}{a real number}
}
\value{
3/4(1-x^2) if x <= 1, 0 otherwise
}
\description{
This function computes the Epanechnikov kernel.
}
\details{
This function computes the Epanechnikov kernel.
}
