% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{norma}
\alias{norma}
\title{Operator squared norm distance}
\usage{
norma(gammax, gamma0)
}
\arguments{
\item{gammax}{the matrix representation of the first operator}

\item{gamma0}{the matrix representation of the second operator}
}
\value{
the squared Frobenius norm of the difference of
the matrices, divided by the number of
elements.
}
\description{
This function computes the squared "operator
norm distance" between its arguments. This is
closely related to the squared Frobenius norm
of the difference of the matrices.
}
