% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{subsets}
\alias{subsets}
\title{All possible subsets}
\usage{
subsets(n, k, set = 1:n)
}
\arguments{
\item{n}{the size of the set of which subsets will be computed}

\item{k}{the size of the subsets}

\item{set}{an optional vector to be taken as the set}
}
\value{
A matrix with subsets listed in its rows
}
\description{
This function returns all possible subsets of a given set.
}
\details{
This function returns all possible subsets of a given set.
}
