% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{tm}
\alias{tm}
\title{Trimmed mean of squares}
\usage{
tm(x, alpha)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{real number between 0 and 1}
}
\value{
the mean of the smallest (1-alpha)% squared values in \code{x}
}
\description{
This function returns the mean of the smallest
(1-alpha)% squares of its arguments.
}
